pyproject.toml
[project]
name = "code-exe-agent"
version = "0.1.0"
description = "MCP Code Execution Structure Generator with AI-powered code generation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    # MCP SDK
    "mcp>=1.0.0",

    # AI/LLM
    "anthropic>=0.40.0",

    # 템플릿 엔진
    "jinja2>=3.1.0",

    # JSON 처리
    "pydantic>=2.0.0",

    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",

    # 유틸리티
    "python-dotenv>=1.0.0",

    # Web UI
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
codex = "main:cli"
codex-agent = "agent_cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
